update quiz_multiple_choice_*
Because of how ActiveRecord works, keeping a reference to the quiz option
caused it to be saved first, thus altering the order of the multiple choice
problems so that the solution was always first.

This gets rid of the table altogether, storing it instead in an array that
gets serialized.